projects/01_fyyur/starter_code/requirements.txt
babel>=2.9.0
python-dateutil>=2.6.0
flask-moment>=0.11.0
flask-wtf>=0.14.3
flask_sqlalchemy>=2.4.4
jinja2==3.0
werkzeug>=2.0.0